adds support for artefact variable category elements
