+++ src/catharsys/api/products/cls_product_view.py
@@ -21,7 +21,6 @@
 ###
 
 import copy
-import enum
 from collections.abc import Iterable
 from pathlib import Path
 from typing import Optional, Callable, Union, Any
@@ -33,167 +32,7 @@ from catharsys.api.products.cls_node import CNode
 from catharsys.api.products.cls_category import CCategory
 
 from .cls_view_dim import CViewDim, CViewDimArtCommon, CViewDimArt, CViewDimArtType, CViewDimGrp, EViewDimType
-
-
-# ##################################################################################################################
-# ##################################################################################################################
-class CViewDimNode:
-    def __init__(self, *, _xProdView: "CProductView", _iDimIdx: int, _sArtTypeId: Optional[str] = None):
-        self._xProdView: "CProductView" = _xProdView
-        self._iDimIdx: int = _iDimIdx
-        self._sArtTypeId: str = _sArtTypeId
-        self._xViewDim: CViewDim = None
-        if self._sArtTypeId is None:
-            self._xViewDim = self._xProdView.lViewDims[self._iDimIdx]
-        else:
-            self._xViewDim = self._xProdView.dicArtViewDims[self._sArtTypeId][self._iDimIdx]
-        # endif
-
-    # enddef
-
-    @property
-    def xViewDim(self) -> CViewDim:
-        return self._xViewDim
-
-    # enddef
-
-    @property
-    def sLabel(self) -> str:
-        return self._xProdView.GetViewDimLabel(self._xViewDim)
-
-    # enddef
-
-    @property
-    def sVarId(self) -> str:
-        return self._xProdView.GetViewDimVarId(self._xViewDim)
-
-    # enddef
-
-    @property
-    def dicCategories(self) -> dict[str, Any]:
-        return self._xProdView.GetViewDimCategories(self._xViewDim)
-
-    # enddef
-
-    @property
-    def sDimLabel(self) -> str:
-        return self._xViewDim.sDimLabel
-
-    # enddef
-
-    @property
-    def sValue(self) -> str:
-        return self._xProdView.GetViewDimValue(self._xViewDim)
-
-    # enddef
-
-    @property
-    def iRange(self) -> int:
-        return self._xViewDim.iRange
-
-    # enddef
-
-    @property
-    def iDimIdx(self) -> int:
-        return self._iDimIdx
-
-    # enddef
-
-    @property
-    def sArtTypeId(self) -> str:
-        return self._sArtTypeId
-
-    # enddef
-
-    @property
-    def bIsUniqueArtVar(self) -> bool:
-        return self._sArtTypeId is not None
-
-    # enddef
-
-    @property
-    def bIsUniqueArtVarStartNode(self) -> bool:
-        return self._sArtTypeId is not None and self._iDimIdx == 0
-
-    # enddef
-
-    @property
-    def lLabels(self) -> Iterable[str]:
-        self._xViewDim.Reset()
-        while True:
-            yield self.sLabel
-            if self._xViewDim.Next() is False:
-                break
-            # endif
-        # endwhile
-
-    # enddef
-
-    @property
-    def lValues(self) -> Iterable[str]:
-        self._xViewDim.Reset()
-        while True:
-            yield self.sValue
-            if self._xViewDim.Next() is False:
-                break
-            # endif
-        # endwhile
-
-    # enddef
-
-    @property
-    def lCategories(self) -> Iterable[dict[str, Any]]:
-        self._xViewDim.Reset()
-        while True:
-            yield self.dicCategories
-            if self._xViewDim.Next() is False:
-                break
-            # endif
-        # endwhile
-
-    # enddef
-
-    def Reset(self):
-        self._xViewDim.Reset()
-
-    # enddef
-
-    def Next(self):
-        return self._xViewDim.Next()
-
-    # enddef
-
-    def NextDim(self) -> Optional["CViewDimNode"]:
-        iNextDim = self._iDimIdx + 1
-        if self._sArtTypeId is None:
-            if iNextDim < len(self._xProdView.lViewDims):
-                return CViewDimNode(_xProdView=self._xProdView, _iDimIdx=iNextDim)
-            # endif
-
-            sViewArtTypeId = self._xProdView.GetViewArtTypeId(_bDoRaise=False)
-            if sViewArtTypeId is None:
-                return None
-            # endif
-
-            lViewDims = self._xProdView.dicArtViewDims.get(sViewArtTypeId)
-            if lViewDims is None or len(lViewDims) == 0:
-                return None
-            # endif
-
-            return CViewDimNode(_xProdView=self._xProdView, _iDimIdx=0, _sArtTypeId=sViewArtTypeId)
-        # endif
-
-        lViewDims = self._xProdView.dicArtViewDims[self._sArtTypeId]
-        if iNextDim < len(lViewDims):
-            return CViewDimNode(_xProdView=self._xProdView, _iDimIdx=iNextDim, _sArtTypeId=self._sArtTypeId)
-        # endif
-
-        return None
-
-    # enddef
-
-
-# endclass
+from .cls_view_dim_node import CViewDimNode
 
 
 # ##################################################################################################################
@@ -243,6 +82,7 @@ class CProductView:
 
         self._dicSelArtVarValueLists: dict[str, list[list[str]]] = dict()
         self._dicSelArtVarLabelLists: dict[str, list[list[str]]] = dict()
+        self._dicSelArtVarCategoryLists: dict[str, list[list[dict[str, Any]]]] = dict()
 
         self._dicSelActArtVarIds: dict[str, list[str]] = dict()
         self._dicSelActArtVarNames: dict[str, list[str]] = dict()
@@ -633,6 +473,7 @@ class CProductView:
         # Get artefact values for selected group values
         self._dicArtVarValueLists, self._dicArtVarTypeLists = self._xProdGrp.GetArtefactVarValues(_lSelGrpVarValueLists)
         self._dicArtVarLabelLists = self._xProdGrp.GetArtefactVarLabels(self._dicArtVarValueLists)
+        self._dicArtVarCategoryLists = self._xProdGrp.GetArtefactVarCategories(self._dicArtVarValueLists)
 
         # List of group variable ids where more than one value is selected.
         # These are the variables that we can iterate over.
@@ -676,6 +517,7 @@ class CProductView:
     def ClearArtefactVarSelection(self):
         self._dicSelArtVarValueLists = dict()
         self._dicSelArtVarLabelLists = dict()
+        self._dicSelArtVarCategoryLists = dict()
         self._dicSelActArtVarIds = dict()
         self._dicSelActArtVarNames = dict()
         self._lSelActArtTypeIds = []
@@ -694,6 +536,12 @@ class CProductView:
         self._dicSelArtVarLabelLists[_sArtTypeId] = self._GetLabelsForSelValues(
             _lSelArtVarValueLists, self._dicArtVarValueLists[_sArtTypeId], self._dicArtVarLabelLists[_sArtTypeId]
         )
+
+        self._dicSelArtVarCategoryLists[_sArtTypeId] = self._GetCategoriesForSelValues(
+            _lSelArtVarValueLists, self._dicArtVarValueLists[_sArtTypeId], self._dicArtVarCategoryLists[_sArtTypeId]
+        )
+
+        # print(f"self._dicSelArtVarCategoryLists: {self._dicSelArtVarCategoryLists}")
 
         # List of artefact variable ids where more than one value is selected.
         # These are the variables that we can iterate over.
@@ -883,6 +731,7 @@ class CProductView:
             lArtVarIdx: list[int] = []
             lArtVarValues: list[str] = None
             lArtVarLabels: list[str] = None
+            lArtVarCatergories: list[dict[str, Any]] = None
             for sArtTypeId in lArtTypeIds:
                 xArtType: CArtefactType = self._xProdGrp.dicArtTypes[sArtTypeId]
                 lArtVarIds = xArtType.xPathStruct.lPathVarIds
@@ -891,6 +740,7 @@ class CProductView:
                 if lArtVarValues is None:
                     lArtVarValues = self._dicSelArtVarValueLists[sArtTypeId][iArtVarIdx]
                     lArtVarLabels = self._dicSelArtVarLabelLists[sArtTypeId][iArtVarIdx]
+                    lArtVarCatergories = self._dicSelArtVarCategoryLists[sArtTypeId][iArtVarIdx]
                 # endif
             # endfor
             if bRangeValid is False:
@@ -901,6 +751,7 @@ class CProductView:
                 _sVarId=sArtVarId,
                 _lValues=lArtVarValues,
                 _lLabels=lArtVarLabels,
+                _lCategories=lArtVarCatergories,
                 _lArtTypeIds=lArtTypeIds,
                 _lVarIdx=lArtVarIdx,
                 _iMin=iMin,
@@ -920,6 +771,7 @@ class CProductView:
             iArtVarIdx = lArtVarIds.index(sArtVarId)
             lArtVarValues = self._dicSelArtVarValueLists[_sArtTypeId][iArtVarIdx]
             lArtVarLabels = self._dicSelArtVarLabelLists[_sArtTypeId][iArtVarIdx]
+            lArtVarCatergories = self._dicSelArtVarCategoryLists[_sArtTypeId][iArtVarIdx]
             if bRangeValid is False:
                 iMin = 0
                 iMax = len(lArtVarValues) - 1
@@ -931,6 +783,7 @@ class CProductView:
                 _iVarIdx=iArtVarIdx,
                 _lValues=lArtVarValues,
                 _lLabels=lArtVarLabels,
+                _lCategories=lArtVarCatergories,
                 _iMin=iMin,
                 _iMax=iMax,
                 _sDimLabel=sDimLabel,
@@ -1171,10 +1024,12 @@ class CProductView:
             dicValue = dict()
 
         elif isinstance(_xViewDim, CViewDimArtCommon):
-            dicValue = dict()
+            xVdac: CViewDimArtCommon = _xViewDim
+            dicValue = xVdac.dicCategories
 
         elif isinstance(_xViewDim, CViewDimArt):
-            dicValue = dict()
+            xVda: CViewDimArt = _xViewDim
+            dicValue = xVda.dicCategories
 
         else:
             raise RuntimeError(f"Invalid view dimension object type: {_xViewDim._eType}")
